pyproject.toml
[tool.poetry]
name = "reliable-api-service"
version = "0.0.0"
description = ""
authors = ["Rees Davison <rees.davison@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"
pydantic = "^2.7.3"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"
pytest-postgresql = "^6.0.0"
pytest-alembic = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
